Factus-backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
